auto-study/requirements.txt
selenium==3.141.0
Pillow==8.0.0
PyQt5==5.15.1
